refactor: migrate field-errors
