+++ packages/components/field-errors/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './field-errors';
+
+export { default as version } from './version';
