--- src/components/field-errors/index.js
@@ -1,1 +1,0 @@
-export { default } from './field-errors';
